requirements.txt
alembic==1.4.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
blueprint==3.4.2
cffi==1.12.3
click==7.1.1
configparser==3.7.4
contextlib2==0.5.5
cryptography==2.9.2
dominate==2.5.1
Faker==4.0.3
Flask==1.1.1
Flask-Admin==1.5.6
flask-blueprint==1.3.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
packaging==20.4
passlib==1.7.2
Pillow==7.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.15
text-unidecode==1.3
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.0
WTForms==2.2.1
